updated instruments routes with instrumentsSetUp
